feat: BeanFactory 초기화 이동
